issue-71: Rename entities to model
