+++ src/MAKE_ATR
@@ -13,14 +13,14 @@
 
 # --- Variables and Configuration --------------------------------------------
 # Folders
-ENTITIES_FOLDER = entities
+MODEL_FOLDER = model
 RASTER_FOLDER = raster
 TESTS_FOLDER = tests
-TESTS_ENTITIES_FOLDER = $(TESTS_FOLDER)\$(ENTITIES_FOLDER)
+TESTS_MODEL_FOLDER = $(TESTS_FOLDER)\$(MODEL_FOLDER)
 TESTS_RASTER_FOLDER = $(TESTS_FOLDER)\$(RASTER_FOLDER)
 UNITY_FOLDER = $(TESTS_FOLDER)\unity
 
-# File Names, Base - Entities
+# File Names, Base - Model
 BLOCK  = block
 SPIKE  = spike
 LAVA   = lava
@@ -54,22 +54,22 @@ ATARI_CC     = cc68x
 SHARED_FLAGS = -g
 
 # Program Configuration
-TOS_FLAGS      = $(SHARED_FLAGS) -c -I. -I$(ENTITIES_FOLDER) -I$(RASTER_FOLDER)
+TOS_FLAGS      = $(SHARED_FLAGS) -c -I. -I$(MODEL_FOLDER) -I$(RASTER_FOLDER)
 RASTER_OBJECTS = $(CLR_SCN).o $(CLR_REG).o $(PLT_PXL).o $(PLT_HOR).o $(PLT_VRT).o $(PLT_LIN).o $(PLT_REC).o $(PLT_SQR).o $(PLT_TRI).o $(BMAP_8).o $(BMAP_16).o $(BMAP_32).o $(PLT_CHR).o $(PLT_STR).o $(BOUNDS).o $(PLT_CLIP).o
 TOS_OBJECTS    = $(BLOCK).o $(SPIKE).o $(LAVA).o $(GEO).o $(LEVEL).o $(WORLD).o $(MODEL).o $(CAMERA).o $(MAIN).o $(RASTER_OBJECTS)
 TOS_NAME       = prog.tos
-MAIN_FLAGS     = $(SHARED_FLAGS) -I. -I$(ENTITIES_FOLDER) -I$(RASTER_FOLDER)
+MAIN_FLAGS     = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(RASTER_FOLDER)
 
 # Test Configuration
 # TODO: It's a bit silly, but pretty sure we're including all our .c files for each test.
-TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(ENTITIES_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
-TEST_SRC   = $(ENTITIES_FOLDER)\$(BLOCK).c $(ENTITIES_FOLDER)\$(SPIKE).c $(ENTITIES_FOLDER)\$(LAVA).c $(ENTITIES_FOLDER)\$(GEO).c $(ENTITIES_FOLDER)\$(LEVEL).c $(ENTITIES_FOLDER)\$(WORLD).c $(ENTITIES_FOLDER)\$(MODEL).c $(ENTITIES_FOLDER)\$(CAMERA).c
+TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(MODEL_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
+TEST_SRC   = $(MODEL_FOLDER)\$(BLOCK).c $(MODEL_FOLDER)\$(SPIKE).c $(MODEL_FOLDER)\$(LAVA).c $(MODEL_FOLDER)\$(GEO).c $(MODEL_FOLDER)\$(LEVEL).c $(MODEL_FOLDER)\$(WORLD).c $(MODEL_FOLDER)\$(MODEL).c $(MODEL_FOLDER)\$(CAMERA).c
 UNITY      = $(UNITY_FOLDER)\unity.c
 
 # Raster Test Configuration
 RASTER_TEST_FLAGS = $(SHARED_FLAGS) -I. -I$(RASTER_FOLDER) -I$(UNITY_FOLDER) -DUNITY_EXCLUDE_STDINT_H -DUNITY_INT_WIDTH=16 -DUNITY_LONG_WIDTH=32 -DUNITY_POINTER_WIDTH=32 -DTESTING
 
-# File Names, Tests - Entities
+# File Names, Tests - Model
 TBLOCK = t${BLOCK}
 TGEO   = t${GEO}
 TLAVA  = t${LAVA}
@@ -101,55 +101,55 @@ TPLT_CLIP = t${PLT_CLIP}
 $(TOS_NAME): $(TOS_OBJECTS)
 	$(ATARI_CC) $(MAIN_FLAGS) $(TOS_OBJECTS) -o $(TOS_NAME)
 
-$(CAMERA).o: $(ENTITIES_FOLDER)\$(CAMERA).c $(ENTITIES_FOLDER)\$(CAMERA).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(CAMERA).c -o $(CAMERA).o
+$(CAMERA).o: $(MODEL_FOLDER)\$(CAMERA).c $(MODEL_FOLDER)\$(CAMERA).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(CAMERA).c -o $(CAMERA).o
 
-$(BLOCK).o: $(ENTITIES_FOLDER)\$(BLOCK).c $(ENTITIES_FOLDER)\$(BLOCK).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(BLOCK).c -o $(BLOCK).o
+$(BLOCK).o: $(MODEL_FOLDER)\$(BLOCK).c $(MODEL_FOLDER)\$(BLOCK).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(BLOCK).c -o $(BLOCK).o
 
-$(LAVA).o: $(ENTITIES_FOLDER)\$(LAVA).c $(ENTITIES_FOLDER)\$(LAVA).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(LAVA).c -o $(LAVA).o
+$(LAVA).o: $(MODEL_FOLDER)\$(LAVA).c $(MODEL_FOLDER)\$(LAVA).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(LAVA).c -o $(LAVA).o
 
-$(SPIKE).o: $(ENTITIES_FOLDER)\$(SPIKE).c $(ENTITIES_FOLDER)\$(SPIKE).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(SPIKE).c -o $(SPIKE).o
+$(SPIKE).o: $(MODEL_FOLDER)\$(SPIKE).c $(MODEL_FOLDER)\$(SPIKE).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(SPIKE).c -o $(SPIKE).o
 
-$(GEO).o: $(ENTITIES_FOLDER)\$(GEO).c $(ENTITIES_FOLDER)\$(GEO).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(GEO).c -o $(GEO).o
+$(GEO).o: $(MODEL_FOLDER)\$(GEO).c $(MODEL_FOLDER)\$(GEO).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(GEO).c -o $(GEO).o
 
-$(LEVEL).o: $(ENTITIES_FOLDER)\$(LEVEL).c $(ENTITIES_FOLDER)\$(LEVEL).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(LEVEL).c -o $(LEVEL).o
+$(LEVEL).o: $(MODEL_FOLDER)\$(LEVEL).c $(MODEL_FOLDER)\$(LEVEL).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(LEVEL).c -o $(LEVEL).o
 
-$(WORLD).o: $(ENTITIES_FOLDER)\$(WORLD).c $(ENTITIES_FOLDER)\$(WORLD).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(WORLD).c -o $(WORLD).o
+$(WORLD).o: $(MODEL_FOLDER)\$(WORLD).c $(MODEL_FOLDER)\$(WORLD).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(WORLD).c -o $(WORLD).o
 
-$(MODEL).o: $(ENTITIES_FOLDER)\$(MODEL).c $(ENTITIES_FOLDER)\$(MODEL).h $(ENTITIES_FOLDER)\$(WORLD).h $(ENTITIES_FOLDER)\dtypes.h
-	$(ATARI_CC) $(TOS_FLAGS) $(ENTITIES_FOLDER)\$(MODEL).c -o $(MODEL).o
+$(MODEL).o: $(MODEL_FOLDER)\$(MODEL).c $(MODEL_FOLDER)\$(MODEL).h $(MODEL_FOLDER)\$(WORLD).h $(MODEL_FOLDER)\dtypes.h
+	$(ATARI_CC) $(TOS_FLAGS) $(MODEL_FOLDER)\$(MODEL).c -o $(MODEL).o
 
-$(MAIN).o: $(MAIN).c $(ENTITIES_FOLDER)\$(MODEL).h $(ENTITIES_FOLDER)\dtypes.h raster\raster.h
+$(MAIN).o: $(MAIN).c $(MODEL_FOLDER)\$(MODEL).h $(MODEL_FOLDER)\dtypes.h raster\raster.h
 	$(ATARI_CC) $(TOS_FLAGS) $(MAIN).c -o $(MAIN).o
 
 
 # ----- Test Builds ----------------------------------------------------------
 $(TBLOCK):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TBLOCK).c $(TEST_SRC) $(UNITY) -o $(TBLOCK).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TBLOCK).c $(TEST_SRC) $(UNITY) -o $(TBLOCK).ut
 
 $(TGEO):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TGEO).c $(TEST_SRC) $(UNITY) -o $(TGEO).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TGEO).c $(TEST_SRC) $(UNITY) -o $(TGEO).ut
 
 $(TLAVA):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TLAVA).c $(TEST_SRC) $(UNITY) -o $(TLAVA).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TLAVA).c $(TEST_SRC) $(UNITY) -o $(TLAVA).ut
 
 $(TLEVEL):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TLEVEL).c $(TEST_SRC) $(UNITY) -o $(TLEVEL).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TLEVEL).c $(TEST_SRC) $(UNITY) -o $(TLEVEL).ut
 
 $(TSPIKE):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TSPIKE).c $(TEST_SRC) $(UNITY) -o $(TSPIKE).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TSPIKE).c $(TEST_SRC) $(UNITY) -o $(TSPIKE).ut
 
 $(TWORLD):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TWORLD).c $(TEST_SRC) $(UNITY) -o $(TWORLD).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TWORLD).c $(TEST_SRC) $(UNITY) -o $(TWORLD).ut
 
 $(TCAMERA):
-	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_ENTITIES_FOLDER)\$(TCAMERA).c $(TEST_SRC) $(UNITY) -o $(TCAMERA).ut
+	$(ATARI_CC) $(TEST_FLAGS) $(TESTS_MODEL_FOLDER)\$(TCAMERA).c $(TEST_SRC) $(UNITY) -o $(TCAMERA).ut
 
 
 # TODO: For some of these, there is cascading dependencies on objects, eg chr
