+++ src/model/camera.c
