+++ src/model/camera.h
