+++ src/model/spike.c
