+++ src/model/spike.h
