+++ src/tests/model/tcamera.c
